% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifold.R
\name{manifold_reduction}
\alias{manifold_reduction}
\title{Compute Optimal Manifold Representation of Points}
\usage{
manifold_reduction(xcoords, no_iterations = 45L, Verbose = TRUE,
  solvemethod = 0L, maxDim = 1.2, knntouse = 75L)
}
\arguments{
\item{xcoords}{d x N matrix of d-dimensional data points}

\item{no_iterations}{The number of iterations to use}

\item{Verbose}{Whether to print status messages}

\item{solvemethod}{An integer from -1 to 5 determining the solver used to
compute the local dimensionality.}

\item{maxDim}{The local dimensionality below which points are fixed. The
default value of \code{1.2} enforces a quasi-1D manifold.}

\item{knntouse}{Number of nearest neighbours to consider when calculating
  interactions.}
}
\value{
a list with the following elements: \itemize{

  \item gamma the d x N coordinates of the output points on manifold

  \item P

  \item lamba determines the tradeoff F(M,Pm) = D + lambda*I

  \item dimension the local dimensionality of the manifold

  }
}
\description{
This implements the algorithm described in Optimal Manifold
  Representation of Data: An Information Theoretic Approach Denis Chigirev
  and William Bialek which attempts to reduce higher dimensional data onto a
  lower dimensional manifold (by default 1D).
}
\references{
\href{http://papers.nips.cc/paper/2399-optimal-manifold-representation-of-data-an-information-theoretic-approach.pdf}{Optimal
Manifold Representation of Data: An Information Theoretic Approach}
}

