% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold-package.R
\docType{package}
\name{manifoldreduction-package}
\alias{manifoldreduction-package}
\alias{manifoldreduction}
\title{Manifold dimension reduction}
\description{
See \code{\link{manifold_reduction}}
}
\seealso{
\code{\link{manifold_reduction}}
}
\keyword{package}
